+++ src/RoadCaptain.Runner/ViewModels/InGameNavigationWindowViewModel.cs
@@ -122,7 +122,7 @@ namespace RoadCaptain.Runner.ViewModels
                 var sportActivity = GetActivityFromSport();
                 Model.UserIsInGame = false;
                 Model.WaitingReason = "Connected with Zwift";
-                Model.InstructionText = $"Start Zwift and start {sportActivity} in {Model.Route.World} on route:";
+                Model.InstructionText = $"Start Zwift and start {sportActivity} in {Model.Route.World.Name} on route:";
             }
             else if (gameState is WaitingForConnectionState && _previousState is InGameState)
             {
@@ -135,7 +135,7 @@ namespace RoadCaptain.Runner.ViewModels
                 var sportActivity = GetActivityFromSport();
                 Model.UserIsInGame = false;
                 Model.WaitingReason = "Waiting for Zwift...";
-                Model.InstructionText = $"Start Zwift and start {sportActivity} in {Model.Route.World} on route:";
+                Model.InstructionText = $"Start Zwift and start {sportActivity} in {Model.Route.World.Name} on route:";
             }
             else if (gameState is ErrorState)
             {